+++ app/src/main/java/com/iyanuoluwa/imovie/DetailsActivity.kt
@@ -8,9 +8,10 @@ import androidx.appcompat.app.AppCompatActivity
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
 import com.bumptech.glide.request.RequestOptions
-import com.iyanuoluwa.imovie.ui.adapter.CastAdapter
-import com.iyanuoluwa.imovie.data.model.buffer.api.Credits
-import com.iyanuoluwa.imovie.data.remote.request.ApiCredits
+import com.iyanuoluwa.imovie.ui.main.BASE_URL
+import com.iyanuoluwa.imovie.ui.details.CastAdapter
+import com.iyanuoluwa.imovie.data.model.Credits
+import com.iyanuoluwa.imovie.data.remote.ApiCredits
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
