+++ app/src/main/java/com/iyanuoluwa/imovie/data/model/Credits.kt
@@ -1,4 +1,4 @@
-package com.iyanuoluwa.imovie.data.model.buffer.api
+package com.iyanuoluwa.imovie.data.model
 
 
 data class Credits(
